Master/requirements.txt
redis
flask
pyzmq